+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed joining through a sign while hotbar slot for leave-item is active
 * Fixed Cage doesn't disappear on first round
 * Added error message if cage floor contains Air
+* Added customizable item name in arena selector gui (by ajgeiss0702)
 
 ### 1.0.5 Release (05.01.2021 - 12.01.2021)
 * Fixed Teleportation to the end is now no longer possible